requirements.txt
jax==0.4.23
jaxlib==0.4.23
scipy>=0.19
numpyro>=0.7.0
flowjax>=0.0.1
equinox<=0.9.0
matplotlib>=3.3.4
arviz>=0.11.2
elfi>=0.8.0
blackjax>=1.0.0
torch>=2.4.0
sbi>=0.23.0
rsnl>=0.0.1
etas>=0.0.1
# latex>=0.7.0
# diffrax>=0.0.1
# psutil>=5.8.0
